Moved cf_connector to libs
